+++ core/components/romanesco/docs/changelog.md
@@ -4,6 +4,7 @@
 Released on ...
 
 Fixes and improvements:
+- Modify getElementDescription snippet to get other fields or property values
 - Add raw placeholder chunks for tag and label
 - Change HTML tag of footer CTA from footer to aside
 - Load SUI form component asynchronous if critical CSS is enabled